+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.8
+
+* Bump Android to 0.7.5
+* Bump iOS to 0.5.18
+* Added `German` and `French` support.
+* Added `enableAutoCapture` to `IdentityBuilder`.
+* Added `validationIncludeList` and `validationExcludeList` to `IdentityBuilder`.
+
 ## 0.2.7
 
 * Bump Android to 0.7.3